backend/requirements.txt
# Core API Framework
fastapi==0.115.5
uvicorn[standard]==0.38.0
pydantic==2.12.4
pydantic-settings==2.6.1
python-dotenv==1.2.1

# Google Cloud & AI
google-cloud-aiplatform==1.73.0
google-auth==2.43.0
google-api-python-client==2.187.0
gspread==6.1.4
gspread-dataframe==4.1.0

# Data Processing
requests==2.32.5
python-dateutil==2.9.0.post0

# Logging & Monitoring
gunicorn==23.0.0
